# docs/swagger.yaml
definitions:
  controller.Class:
    properties:
      course_title:
        type: string
      metadata: {}
      title:
        type: string
    type: object
  controller.CourseRequest:
    properties:
      category:
        type: string
      classes:
        items:
          $ref: '#/definitions/controller.Class'
        type: array
      is_school_oriented:
        type: boolean
      max_age:
        type: integer
      metadata: {}
      min_age:
        type: integer
      title:
        type: string
    type: object
  controller.CourseState:
    properties:
      age_filtered:
        type: boolean
      category:
        type: string
      classes:
        items:
          type: string
        type: array
      courseTitle:
        type: string
      creatorEmail:
        type: string
      isSchoolOriented:
        type: boolean
      isSubscribed:
        type: boolean
      max_age:
        type: integer
      metadata: {}
      min_age:
        type: integer
    type: object
  controller.CourseStateResponse:
    properties:
      amount:
        type: integer
      courses:
        items:
          $ref: '#/definitions/controller.CourseState'
        type: array
    type: object
...
  controller.ErrorMsg:
    properties:
      reason:
        example: mensaje de error
        type: string
    type: object
...
  controller.SubscriptionRequest:
    properties:
      course_title:
        type: string
      metadata: {}
      user_id:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /course/:
    get:
      consumes:
      - application/json
      description: Get all courses that follows a criteria
      parameters:
      - description: Title string for which you want to look
        in: query
        name: title
        type: string
      - description: ownerEmail string for which you want to look
        in: query
        name: ownerEmail
        type: string
      - description: category string for which you want to look
        in: query
        name: category
        type: string
      - description: Age of the course you want to retrieve
        in: query
        name: desiredAge
        type: integer
      - description: true if school oriented, any other value otherwise
        in: query
        name: isSchoolOriented
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CourseStateResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorMsg'
      summary: Get all courses
      tags:
      - Course request
    post:
      consumes:
      - application/json
      description: Create course using the token as a way to add account to course
        owner
      parameters:
      - description: Title and Category are required
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/controller.CourseRequest'
      - description: token required for request
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CourseState'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorMsg'
      summary: Create course
      tags:
      - Course request
  /course/{id}:
    get:
      consumes:
      - application/json
      description: Fetch a course with a given id
      parameters:
      - description: course id which you look for
        in: path
        name: id
        required: true
        type: string
      - description: token required for request
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CourseState'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorMsg'
      summary: Fetch a course
      tags:
      - Course request
    post:
      consumes:
      - application/json
      description: Create class for a previously created course, if course does not
        exist this endpoint will fail
      parameters:
      - description: Title is required
        in: body
        name: class
        required: true
        schema:
          $ref: '#/definitions/controller.Class'
      - description: course id which you want to add a course
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CourseState'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorMsg'
      summary: Create class for created course
      tags:
      - Course request
  /course/{id}/{classId}:
    delete:
      consumes:
      - application/json
      description: Removes a class
      parameters:
      - description: class id you want to remove
        in: path
        name: classId
        required: true
        type: string
      - description: course id which you look for
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorMsg'
      summary: Remove class created
      tags:
      - Course request
    get:
      consumes:
      - application/json
      description: Get class with id and class id
      parameters:
      - description: class id you want to fetch
        in: path
        name: classId
        required: true
        type: string
      - description: course id which you look for
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Class'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorMsg'
      summary: Fetch a class
      tags:
      - Course request
  /course/subscribe/:
    get:
      consumes:
      - application/json
      description: Get all courses in which the user has subscribed
      parameters:
      - description: token required for request
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CourseStateResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorMsg'
      summary: Get subscribed courses
      tags:
      - Subscription
  /course/subscribe/{id}:
    post:
      consumes:
      - application/json
      description: Subscribe a user given by its token to a course
      parameters:
      - description: course in which the current user wants to subscribe
        in: path
        name: id
        required: true
        type: string
      - description: token required for request
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: Object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorMsg'
      summary: Subscribe
      tags:
      - Subscription
  /course/subscribe/courses/{id}:
    get:
      consumes:
      - application/json
      description: Get all courses in which the user has subscribed
      parameters:
      - description: course id for which you want to get user subscribed
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controller.SubscriptionRequest'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorMsg'
      summary: Get subscribed courses
      tags:
      - Subscription
  /courses/:
    get:
      consumes:
      - application/json
      description: This is just a wrapper of GetCourses, in which you send the token,
        and we return the all courses
      parameters:
      - description: token required for request
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controller.CourseStateResponse'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorMsg'
      summary: Get all own courses
      tags:
      - Course request
...
swagger: "2.0"